first draft for control panel. Needs a QUESTION/LEVEL img, has coins, rubies, beans and level/questions and holds the newsfeed.
