+++ src/components/controlpanel/control.css
@@ -1,0 +1,54 @@
+#control-panel {
+  position: absolute;
+  width: 295px;
+  height: 100%;
+  right: 0;
+  top: 0;
+  border-left: 5px solid darkgoldenrod
+}
+
+/* STATUS BAR */
+
+#player-status-bar {
+  text-align: center;
+  color: white;
+}
+
+.status-bar-image {
+  background-size: 30px;
+  width: 50px;
+  height: 29px;
+  text-align: right;
+  background-position: left;
+}
+
+.cash-status {
+  background-size: 20px;
+  width: 60px;
+  height: 29px;
+  text-align: right;
+  background-position: left;
+}
+
+.status-count {
+  text-align: center;
+  width: 73px;
+  height: 40px;
+  float: right;
+  font-size: 1.4em;
+}
+
+/* NEWSFEED SECTION */
+
+#newsfeed-assemble {
+  position: absolute;
+  text-align: center;
+  color: white;
+  height: 150px;
+  width: 95%;
+  top: 100px;
+}
+
+.newsfeed-message {
+  padding: 2px;
+}
