+++ src/index.css
@@ -23,15 +23,6 @@ body {
   border: 4px solid darkgoldenrod;
 }
 
-.newsfeed-assemble {
-  text-align: center;
-  color: white;
-}
-
-.newsfeed-message {
-  padding: 2px;
-}
-
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
@@ -156,11 +147,4 @@ code {
   background-image: url('/assets/signs.png');
   background-repeat: no-repeat;
   background-position: center;
-}
-
-#player-status-bar {
-  display: inline;
-  float: left;
-  color: white
-
 }
